Updates for Linux Fixed ctors and other inconsistencies.
git-svn-id: https://tpb3dtg.svn.sourceforge.net/svnroot/tpb3dtg@16 7c74f4f4-39d4-4077-a54d-10deffb394f4

+++ CoreObjects/RideBase.cpp
@@ -32,7 +32,7 @@ void RideBase::Load(SerializerBase& ser)
 void RideBase::Save(SerializerBase& ser)
 {
    ser.Add("type", "Track");
-   long* pI = (long*)mRGBA[0];
+   int* pI = (int*)mRGBA[0];
    ser.Add("RGBA", *pI);
    ser.Add("ID", mID);
 }
